requirements.txt
Django==1.9.4
djangorestframework==3.3.3
django-filter==0.13.0